+++ test/e2e/tests/identity/contact-syncing/sync-existing-user.spec.ts
@@ -11,7 +11,7 @@ import {
 } from '../constants';
 import { UserStorageMockttpController } from '../../../helpers/identity/user-storage/userStorageMockttpController';
 import { createEncryptedResponse } from '../../../helpers/identity/user-storage/generateEncryptedData';
-import { completeOnboardFlowIdentity } from '../flows';
+import { completeOnboardFlowIdentity } from '../../../page-objects/flows/identity.flow';
 import HeaderNavbar from '../../../page-objects/pages/header-navbar';
 import ContactsSettings from '../../../page-objects/pages/settings/contacts-settings';
